Move to Shared - SqlSer.cs
This is part of Code merge for issue #1261. I have created .common.cs and moved the uncommon code to private methods.
